make resolvers moduled
