+++ api/src/graphql/index.ts
@@ -3,5 +3,14 @@ import userTypeDefs from './users/users.graphql';
 
 import itemsResolvers from './items/items.resolvers';
 import usersResolvers from './users/users.resolvers';
+import followsResolvers from './follows/follows.resolvers';
+import chatsResolvers from './chats/chats.resolvers';
 
-export { itemTypeDefs, userTypeDefs, itemsResolvers, usersResolvers };
+export {
+  itemTypeDefs,
+  userTypeDefs,
+  itemsResolvers,
+  usersResolvers,
+  followsResolvers,
+  chatsResolvers,
+};
