+++ api/src/graphql/items/items.resolvers.ts
@@ -1,8 +1,23 @@
+import { v4 as uuidV4 } from 'uuid';
 import { Op } from 'sequelize';
-import { Item } from '../../models/sequelize';
+import { Item, User, UserItem } from '../../models/sequelize';
+import CreateItemInput from '../../models/ts/CreateItemInput';
 
 export = {
   Query: {
+    item: async (_: any, args: { id: string }) => {
+      const result = await Item.findByPk(args.id);
+      const itemData = result.dataValues;
+
+      return {
+        id: itemData.id,
+        title: itemData.title,
+        author: itemData.author,
+        imageUrl: itemData.imageUrl,
+        users: [],
+      };
+    },
+
     // FIXME: any type to something
     itemsByTitle: async (_: any, args: any) => {
       const titleQuery = args.title;
@@ -22,6 +37,86 @@ export = {
       });
 
       return itemList;
+    },
+
+    fetchRandomItems: async () => {
+      const items = await Item.findAll({
+        limit: 1,
+      });
+
+      const itemList = items.map((elm: any) => ({
+        id: elm.id,
+        title: elm.title,
+        author: elm.author,
+        imageUrl: elm.imageUrl,
+      }));
+
+      return itemList;
+    },
+  },
+  Mutation: {
+    createItem: async (_: any, args: { data: CreateItemInput }) => {
+      const id = uuidV4();
+      const { title, author, imageUrl } = args.data;
+      const newItem = { id, title, author, imageUrl };
+
+      await Item.create(newItem);
+
+      return newItem;
+    },
+
+    deleteUserItem: async (
+      _: any,
+      args: {
+        data: { userId: string; itemId: string };
+      },
+    ) => {
+      const userItemInstance = await UserItem.findOne({
+        where: { userId: args.data.userId, itemId: args.data.itemId },
+      });
+      if (userItemInstance) {
+        await userItemInstance.destroy();
+      }
+
+      // create return value
+      const userInstance = await User.findOne({
+        where: { id: args.data.userId },
+        include: Item,
+      });
+
+      return {
+        id: userInstance.id,
+        name: userInstance.name,
+        about: userInstance.about,
+        imageUrl: userInstance.imageUrl,
+        items: userInstance.items,
+      };
+    },
+
+    addUserItem: async (
+      _: any,
+      args: { data: { userId: string; itemId: string } },
+    ) => {
+      // FIXME: check if user and item exists
+
+      await UserItem.create({
+        userId: args.data.userId,
+        itemId: args.data.itemId,
+      });
+
+      // create return value
+      const userInstance = await User.findOne({
+        where: { id: args.data.userId },
+        include: Item,
+      });
+
+      return {
+        id: userInstance.id,
+        name: userInstance.name,
+        about: userInstance.about,
+        imageUrl: userInstance.imageUrl,
+        items: userInstance.items,
+      };
     },
   },
 };
