+++ api/src/graphql/users/users.resolvers.ts
@@ -1,6 +1,7 @@
 import { QueryTypes } from 'sequelize';
 import { Item, User } from '../../models/sequelize';
 import sequelize from '../../util/database';
+import UserType from '../../models/ts/User';
 
 export = {
   Query: {
@@ -34,6 +35,21 @@ export = {
         imageUrl: userData.imageUrl,
         items,
       };
+    },
+
+    getUserCount: async (_: any, args: any) => {
+      console.log('getUserCount args.id', args.id);
+
+      try {
+        console.log('User.count');
+
+        const amount = await User.count({ where: { id: args.id } });
+        console.log('amount', amount);
+
+        return amount;
+      } catch (err) {
+        console.log('getUserCount error', err);
+      }
     },
 
     getUsersByItems: async (
@@ -92,6 +108,42 @@ export = {
       );
 
       return fetchedUsers;
+    },
+  },
+
+  Mutation: {
+    createUser: (_: any, args: { id: string }) => {
+      const newUser = User.create({
+        id: args.id,
+        name: 'new user',
+        about: ``,
+        imageUrl: '',
+      });
+      return newUser;
+    },
+
+    updateUser: async (_: any, args: { data: UserType }) => {
+      const userInstance = await User.findByPk(args.data.id);
+      if (!userInstance) {
+        throw new Error('User not found!');
+      }
+
+      // update
+      if (args.data && args.data.name) {
+        userInstance.name = args.data.name;
+        userInstance.about = args.data.about;
+        userInstance.imageUrl = args.data.imageUrl;
+
+        // save
+        userInstance.save();
+      }
+
+      return {
+        id: userInstance.id,
+        about: userInstance.about,
+        imageUrl: userInstance.imageUrl,
+        name: userInstance.name,
+      };
     },
   },
 };
